+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## `jupyter_starters 0.1.0a1`
+
+- add more sources of config
+
 ## `@deathbeds/jupyterlab-starters 0.1.0a0`
 
 - initial implementation